# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Project Management System'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/reset-password:
    post:
      summary: 'Reset Password.'
      operationId: resetPassword
      description: ''
      parameters: []
      responses:
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'The password field confirmation does not match.'
                  errors:
                    password:
                      - 'The password field confirmation does not match.'
                properties:
                  message:
                    type: string
                    example: 'The password field confirmation does not match.'
                  errors:
                    type: object
                    properties:
                      password:
                        type: array
                        example:
                          - 'The password field confirmation does not match.'
                        items:
                          type: string
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: ''
                  example: consectetur
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: noah31@example.net
                password:
                  type: string
                  description: 'Must be at least 6 characters.'
                  example: '{;aKHTU!2'
              required:
                - token
                - email
                - password
      security: []
...